// lib/generated/RoutingNode.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RoutingPred } from "./RoutingPred";

export type RoutingNode = { key: string, kind: "STOP" | "BAJS", lat: number, lon: number, name: string, time: number, pred?: RoutingPred, };